requirements.txt
pyautogui>=0.9.54
python-chess>=1.10.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
keyboard>=0.13.5
mss>=9.0.0
websocket-client>=1.6.0
requests>=2.31.0
